refactor: 필요한 메서드만 넘겨주도록 수정
